+++ java-design-processor/src/test/resources/internal/nbbrd/design/TraitNotInterface.java
@@ -1,0 +1,10 @@
+package internal.nbbrd.design;
+
+import nbbrd.design.Trait;
+
+@Trait
+public class TraitNotInterface {
+
+    void move(int x, int y) {
+    }
+}